sdxl_dpo/requirements.txt
fsspec==2024.9.0
packaging==24.2
torchvision==0.20.1
pillow==11.0.0
sentry-sdk==2.18.0
setuptools==75.1.0
aiohttp==3.11.2
wheel==0.44.0
six==1.16.0
attrs==24.2.0
aiosignal==1.3.1
uv==0.5.2
huggingface-hub==0.26.2
protobuf==5.28.3
zipp==3.21.0
nvidia-cuda-cupti-cu12==12.4.127
triton==3.1.0
tzdata==2024.2
nvidia-cublas-cu12==12.4.5.8
safetensors==0.4.5
propcache==0.2.0
datasets==3.1.0
nvidia-nccl-cu12==2.21.5
nvidia-cusolver-cu12==11.6.1.9
python-dateutil==2.9.0.post0
wandb==0.18.7
aiohappyeyeballs==2.4.3
dill==0.3.8
nvidia-cufft-cu12==11.2.1.3
tokenizers==0.20.3
networkx==3.4.2
nvidia-cusparse-cu12==12.3.1.170
charset-normalizer==3.4.0
yarl==1.17.2
smmap==5.0.1
numpy==2.1.3
certifi==2024.8.30
idna==3.10
sympy==1.13.1
nvidia-nvtx-cu12==12.4.127
click==8.1.7
transformers==4.46.2
xformers==0.0.28.post3
docker-pycreds==0.4.0
platformdirs==4.3.6
importlib_metadata==8.5.0
diffusers==0.31.0
multidict==6.1.0
psutil==6.1.0
xxhash==3.5.0
PyYAML==6.0.2
multiprocess==0.70.16
typing_extensions==4.12.2
peft==0.13.2
bitsandbytes==0.44.1
MarkupSafe==3.0.2
nvidia-nvjitlink-cu12==12.4.127
pip==24.2
regex==2024.11.6
mpmath==1.3.0
nvidia-curand-cu12==10.3.5.147
urllib3==2.2.3
filelock==3.16.1
pyarrow==18.0.0
GitPython==3.1.43
Jinja2==3.1.4
nvidia-cuda-nvrtc-cu12==12.4.127
pandas==2.2.3
nvidia-cudnn-cu12==9.1.0.70
setproctitle==1.3.4
pytz==2024.2
frozenlist==1.5.0
torch==2.5.1
nvidia-cuda-runtime-cu12==12.4.127
requests==2.32.3
accelerate==1.1.1
gitdb==4.0.11
tqdm==4.67.0
